Adds Trans variable support
- Can now use variables in Trans components
- Breaks up trans and file specific utils into separate files
