+++ __tests__/fixtures/parent.tsx
@@ -9,12 +9,17 @@ interface Props {
 class MyComponent extends React.Component<Props, any> {
   render() {
     const { t } = this.props;
+    const name = 'Annalise';
+    const day = 'Tuesday';
+    const boys = t('{{count}} boys', { count: 3 });
+    const girls = t('{{count}} girls', { count: 1 });
 
     return (
       <div>
         <h1>Parent</h1>
+        <Trans>{{boys}} and {{girls}}</Trans>
         <Trans>Hello</Trans>
-        <Trans name='Annalise'>Hello {{name}}</Trans>
+        <Trans>Hello {{name}} it's {{day}}</Trans>
         <Child />
       </div>
     );
