+++ src/index.ts
@@ -1,7 +1,8 @@
 export {
   default as i18nJsonLoader,
-  languages,
   parser,
 } from './i18n-json-webpack-loader';
+
+export { languages } from './util/file';
 
 export { default as I18nJsonTransformPlugin } from './i18n-json-transform-webpack-plugin';
